Refactor hackathon management to use Supabase directly
- Updated HackathonPage to use useHackathonById hook for fetching hackathon data.
- Replaced API client with direct Supabase queries for creating, fetching, updating, and deleting hackathons.
- Simplified error handling and loading states in CreateHackathonForm.
- Removed obsolete API client and server database functions.
- Enhanced data insertion logic for related entities (prize cohorts, judges, schedule slots) during hackathon creation and update.
