+++ src/components/hackathon/create/form.tsx
@@ -10,9 +10,8 @@ import {
   validateDateConsistency,
 } from "@/lib/schemas/hackathon-schema";
 import { toast } from "sonner";
-import { useMutation } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
-import { apiClient, ApiError } from "@/lib/api/client";
+import { useCreateHackathon } from "@/hooks/queries/use-hackathons";
 import { getRandomHackathons } from "@/data/hackathons";
 
 export function CreateHackathonForm() {
@@ -107,44 +106,25 @@ export function CreateHackathonForm() {
   const { handleSubmit, setError, clearErrors } = methods;
 
   // Create hackathon mutation
-  const createHackathonMutation = useMutation({
-    mutationFn: (data: HackathonFormData) => apiClient.createHackathon(data),
-    onSuccess: (result) => {
-      toast.success(result.message || "Hackathon created successfully!");
-      router.push("/dashboard");
-    },
-    onError: (error) => {
-      console.error("Error creating hackathon:", error);
+  const createHackathonMutation = useCreateHackathon();
 
-      if (error instanceof ApiError) {
-        // Handle validation errors
-        if (error.status === 400 && error.details) {
-          // Set field errors from validation
-          Object.entries(error.details).forEach(([field, messages]) => {
-            if (Array.isArray(messages) && messages.length > 0) {
-              setError(field as any, {
-                type: "server",
-                message: messages[0] as string,
-              });
-            }
-          });
-          toast.error("Please fix the validation errors");
-          return;
-        }
+  const handleCreateSuccess = () => {
+    toast.success("Hackathon created successfully!");
+    router.push("/dashboard");
+  };
 
-        // Handle authentication errors
-        if (error.status === 401) {
-          toast.error("Please log in to create a hackathon");
-          router.push("/login");
-          return;
-        }
+  const handleCreateError = (error: Error) => {
+    console.error("Error creating hackathon:", error);
 
-        toast.error(error.message);
-      } else {
-        toast.error("Failed to create hackathon");
-      }
-    },
-  });
+    // Handle authentication errors
+    if (error.message.includes("Authentication")) {
+      toast.error("Please log in to create a hackathon");
+      router.push("/login");
+      return;
+    }
+
+    toast.error(error.message || "Failed to create hackathon");
+  };
 
   const onSubmit = async (data: HackathonFormData) => {
     console.log("Form submitted with data:", data);
@@ -163,7 +143,10 @@ export function CreateHackathonForm() {
 
     clearErrors();
     console.log("Creating hackathon...");
-    createHackathonMutation.mutate(data);
+    createHackathonMutation.mutate(data, {
+      onSuccess: handleCreateSuccess,
+      onError: handleCreateError,
+    });
   };
 
   return (
